Extract `printTypeParameter`
